refactor(uniquet): created a subpackage to prevent conflicts
