+++ element-template-generator/uniquet/src/main/java/io/camunda/connector/uniquet/Main.java
@@ -14,11 +14,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.connector.dto;
+package io.camunda.connector.uniquet;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
+import io.camunda.connector.uniquet.command.UniquetCommand;
+import picocli.CommandLine;
 
-@JsonIgnoreProperties(ignoreUnknown = true)
-public record ElementTemplate(
-    @JsonProperty(required = true) String id, @JsonProperty(required = true) Integer version) {}
+public class Main {
+  public static void main(String[] args) {
+    int exitCode =
+        new CommandLine(new UniquetCommand())
+            .setUnmatchedOptionsArePositionalParams(true)
+            .execute(args);
+    System.exit(exitCode);
+  }
+}
